kernel: Remove enhanced security feature (#3123)
This feature does not work with tracepoint hook because setresuid is not
even hooked for apps not in allowlist. It was invented when we still use
task_fix_setuid LSM hook, but since it is inlined in oplus kernel and we
switched to tracepoint hook, we don't have a reliable way to implement
this feature without introducing side channels. Remove it and avoid
confuse users.

Signed-off-by: AlexLiuDev233 <wzylin11@outlook.com>